fix bug 0021552: error while importing data from folder
git-svn-id: http://talendforge.org/svn/tom/trunk@60393 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.mdm.workbench/src/com/amalto/workbench/export/ImportItemsWizard.java
@@ -182,7 +182,8 @@ public class ImportItemsWizard extends Wizard {
     public boolean performFinish() {
         closeOpenEditors();
         final String zipFilePath = getZipFilePath();
-        if (zipBtn.getSelection()) {
+        final boolean selectZip = zipBtn.getSelection();
+        if (selectZip) {
 
             // importFolder= System.getProperty("user.dir")+"/temp";
             try {
@@ -220,8 +221,8 @@ public class ImportItemsWizard extends Wizard {
                         }
 
                     }.schedule();
-
-                    if (zipFilePath != null) {
+                    //modified by honghb ,fix bug 21552
+                    if (selectZip && zipFilePath != null) {
                         importFolder = System.getProperty("user.dir") + "/temp";//$NON-NLS-1$//$NON-NLS-2$
                         ZipToFile.deleteDirectory(new File(importFolder));
                     }
